clase06 terminada
