+++ index.html
@@ -40,6 +40,6 @@
         </ul>
       </div>
     </main>
-    <script src="scripts.js"></script>
+    <script type='module' src="scripts.js"></script>
   </body>
 </html>
